# svc/auth/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: OpenAPI contract for the Go auth service.
servers:
  - url: http://localhost:4000
tags:
  - name: Health
  - name: Auth
  - name: Keys
  - name: Store
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Returns a simple status payload for readiness/liveness checks.
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up a user or OAuth account
      description: Creates an account using email/password or provider token onboarding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in a user or OAuth account
      description: Authenticates with email/password or provider token onboarding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/oauth/{provider}/callback:
    get:
      tags: [Auth]
      summary: OAuth callback
      description: Provider callback endpoint for Google and GitHub.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: token
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OAuth account linked and session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /keys/service-account/create:
    post:
      tags: [Keys]
      summary: Create a service account
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountCreateRequest"
      responses:
        "201":
          description: Service account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/service-account/update:
    put:
      tags: [Keys]
      summary: Update a service account
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountUpdateRequest"
      responses:
        "200":
          description: Service account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/service-account/delete:
    delete:
      tags: [Keys]
      summary: Delete a service account
      security:
        - sessionBearer: []
      parameters:
        - name: uid
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/initiate-auth:
    post:
      tags: [Keys]
      summary: Start service-account authentication
      description: Issues a one-time signing token for the two-step auth flow.
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateAuthRequest"
      responses:
        "200":
          description: Signing token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiateAuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/authenticate:
    post:
      tags: [Keys]
      summary: Finish service-account authentication
      description: Verifies the signing token and returns a short-lived JWT.
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticateRequest"
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /store/get:
    get:
      tags: [Store]
      summary: Proxy a GET request through the RBAC gate
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /store/{proxy}:
    parameters:
      - name: proxy
        in: path
        required: true
        schema:
          type: string
        description: Upstream path segment after /store/.
    get:
      tags: [Store]
      summary: Proxy GET requests
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Store]
      summary: Proxy POST requests
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Store]
      summary: Proxy PUT requests
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags: [Store]
      summary: Proxy PATCH requests
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Store]
      summary: Proxy DELETE requests
      security:
        - jwtBearer: []
      responses:
        "200":
          description: Upstream response
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    jwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: RBAC denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HealthResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        db:
          type: string
          example: up
        redis:
          type: string
          example: up
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        provider:
          type: string
          enum: [google, github]
        token:
          type: string
      additionalProperties: false
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        provider:
          type: string
          enum: [google, github]
        token:
          type: string
      additionalProperties: false
    AuthResponse:
      type: object
      properties:
        uid:
          type: string
        email:
          type: string
        account_type:
          type: string
          enum: [user, team]
        organization:
          type: string
        session_token:
          type: string
        jwt:
          type: string
        provider:
          type: string
        provider_uid:
          type: string
    Scope:
      type: object
      required: [root, path, scope]
      properties:
        root:
          type: string
          enum: [user, team]
        path:
          type: string
        scope:
          type: string
          enum: [read, memory, write]
    ServiceAccountCreateRequest:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
        secret_key:
          type: string
        expiry_date:
          type: string
          description: Date in YYYY-MM-DD format
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/Scope"
      additionalProperties: false
    ServiceAccountUpdateRequest:
      type: object
      required: [uid, name, scopes]
      properties:
        uid:
          type: string
        name:
          type: string
        secret_key:
          type: string
        expiry_date:
          type: string
          description: Date in YYYY-MM-DD format
        is_active:
          type: boolean
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/Scope"
      additionalProperties: false
    ServiceAccountResponse:
      type: object
      properties:
        uid:
          type: string
        access_key:
          type: string
        name:
          type: string
        secret_key:
          type: string
        expiry_date:
          type: string
          format: date-time
        is_active:
          type: boolean
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/Scope"
    InitiateAuthRequest:
      type: object
      required: [service_account_token]
      properties:
        service_account_token:
          type: string
    InitiateAuthResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AuthenticateRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
        signedPayload:
          type: string
    AuthenticateResponse:
      type: object
      required: [jwt]
      properties:
        jwt:
          type: string